feat: export svg, fix menu styles
